+++ __tests__/compilers/compatability.test.tsx
@@ -75,7 +75,7 @@ describe('compatability with RDMD', () => {
     };
 
     expect(mdx(ast).trim()).toMatchInlineSnapshot(`
-      "<Image align="center" width="300px" src="https://drastik.ch/wp-content/uploads/2023/06/blackcat.gif" alt="" title="">
+      "<Image align="center" width="300px" src="https://drastik.ch/wp-content/uploads/2023/06/blackcat.gif" border={true}>
         hello **cat**
       </Image>"
     `);
@@ -205,8 +205,6 @@ This is an image: <img src="http://example.com/#\\>" >
   });
 
   describe('<HTMLBlock> wrapping', () => {
-  // configure({ defaultIgnore: undefined });
-
     const rawStyle = `<style data-testid="style-tag">
     p {
       color: red;
@@ -266,5 +264,242 @@ This is an image: <img src="http://example.com/#\\>" >
       );
       expect(spy).toHaveBeenCalledTimes(0);
     });
+  });
+
+  it('can parse and transform magic image block AST to MDX', () => {
+    const md = `
+[block:image]
+{
+  "images": [
+    {
+      "image": [
+        "https://files.readme.io/4a1c7a0-Iphone.jpeg",
+        null,
+        ""
+      ],
+      "align": "center",
+      "sizing": "250px"
+    }
+  ]
+}
+[/block]
+`;
+
+    const rmdx = mdx(rdmd.mdast(md));
+
+    expect(rmdx).toMatch('<Image align="center" width="250px" src="https://files.readme.io/4a1c7a0-Iphone.jpeg" />');
+  });
+
+  it('can parse and transform magic image block AST to MDX with caption', () => {
+    const md = `
+[block:image]
+{
+  "images": [
+    {
+      "image": [
+        "https://files.readme.io/fd21f977cfbb9f55b3a13ab0b827525e94ee1576f21bbe82945cdc22cc966d82-Screenshot_2024-09-12_at_3.47.05_PM.png",
+        "",
+        "Data Plane Setup"
+      ],
+      "align": "center",
+      "caption": "Data Plane Setup",
+      "border": true
+    }
+  ]
+}
+[/block]`;
+
+    const rmdx = mdx(rdmd.mdast(md));
+    expect(rmdx).toMatchInlineSnapshot(
+      `
+      "<Image alt="Data Plane Setup" align="center" border={true} src="https://files.readme.io/fd21f977cfbb9f55b3a13ab0b827525e94ee1576f21bbe82945cdc22cc966d82-Screenshot_2024-09-12_at_3.47.05_PM.png">
+        Data Plane Setup
+      </Image>
+      "
+    `,
+    );
+  });
+
+  it('trims whitespace surrounding phrasing content (emphasis, strong, etc)', () => {
+    const md = `** bold ** and _ italic _ and ***   bold italic ***`;
+
+    const rmdx = mdx(rdmd.mdast(md));
+    expect(rmdx).toMatchInlineSnapshot(`
+      "**bold** and *italic* and ***bold italic***
+      "
+    `);
+  });
+
+  it('moves whitespace surrounding phrasing content (emphasis, strong, etc) to the appropriate place', () => {
+    const md = `**bold **and also_ italic_ and*** bold italic***aaaaaah`;
+
+    const rmdx = mdx(rdmd.mdast(md));
+    expect(rmdx).toMatchInlineSnapshot(`
+      "**bold** and also *italic* and ***bold italic***aaaaaah
+      "
+    `);
+  });
+
+  it('correctly parses and transforms image magic block with legacy data', () => {
+    const md = `
+[block:image]
+{
+  "images": [
+    {
+      "image": [
+        "https://files.readme.io/9ac3bf4-SAP_Folders_Note.png",
+        "SAP Folders Note.png",
+        806,
+        190,
+        "#e8e9ea"
+      ]
+    }
+  ]
+}
+[/block]
+## Tile View
+    `;
+
+    const tree = rdmd.mdast(md);
+    const rmdx = mdx(tree);
+    expect(rmdx).toMatchInlineSnapshot(`
+      "![806](https://files.readme.io/9ac3bf4-SAP_Folders_Note.png "SAP Folders Note.png")
+
+      ## Tile View
+      "
+    `);
+  });
+
+  it('compiles parameter magic blocks with breaks to jsx', () => {
+    const md = `
+[block:parameters]
+${JSON.stringify(
+  {
+    data: {
+      'h-0': 'Term',
+      'h-1': 'Definition',
+      '0-0': 'Events',
+      '0-1': 'Pseudo-list:  \n● One  \n● Two',
+    },
+    cols: 2,
+    rows: 1,
+    align: ['left', 'left'],
+  },
+  null,
+  2,
+)}
+[/block]
+`;
+
+    const rmdx = mdx(rdmd.mdast(md));
+    expect(rmdx).toMatchInlineSnapshot(`
+      "<Table align={["left","left"]}>
+        <thead>
+          <tr>
+            <th style={{ textAlign: "left" }}>
+              Term
+            </th>
+
+            <th style={{ textAlign: "left" }}>
+              Definition
+            </th>
+          </tr>
+        </thead>
+
+        <tbody>
+          <tr>
+            <td style={{ textAlign: "left" }}>
+              Events
+            </td>
+
+            <td style={{ textAlign: "left" }}>
+              Pseudo-list:
+
+
+
+
+              ● One
+
+
+
+
+              ● Two
+            </td>
+          </tr>
+        </tbody>
+      </Table>
+      "
+    `);
+  });
+
+  it('compiles callouts without a title', () => {
+    const md = `
+> 🥈
+>
+> Lorem ipsum dolor sit amet consectetur adipisicing elit. Error eos animi obcaecati quod repudiandae aliquid nemo veritatis ex, quos delectus minus sit omnis vel dolores libero, recusandae ea dignissimos iure?
+`;
+
+    const rmdx = mdx(rdmd.mdast(md));
+    expect(rmdx).toMatchInlineSnapshot(`
+      "> 🥈
+      >
+      > Lorem ipsum dolor sit amet consectetur adipisicing elit. Error eos animi obcaecati quod repudiandae aliquid nemo veritatis ex, quos delectus minus sit omnis vel dolores libero, recusandae ea dignissimos iure?
+      "
+    `);
+  });
+
+  it('compiles tables with inline code with newlines', () => {
+    const md = `
+[block:parameters]
+${JSON.stringify(
+  {
+    data: {
+      'h-0': 'Field',
+      'h-1': 'Description',
+      '0-0': 'orderby',
+      '0-1': '`{\n "field": "ID",\n "type": "ASC"\n }`',
+    },
+    cols: 2,
+    rows: 1,
+    align: ['left', 'left'],
+  },
+  null,
+  2,
+)}
+[/block]
+    `;
+
+    const rmdx = mdx(rdmd.mdast(md));
+    expect(rmdx).toMatchInlineSnapshot(`
+      "<Table align={["left","left"]}>
+        <thead>
+          <tr>
+            <th style={{ textAlign: "left" }}>
+              Field
+            </th>
+
+            <th style={{ textAlign: "left" }}>
+              Description
+            </th>
+          </tr>
+        </thead>
+
+        <tbody>
+          <tr>
+            <td style={{ textAlign: "left" }}>
+              orderby
+            </td>
+
+            <td style={{ textAlign: "left" }}>
+              \`{
+               "field": "ID",
+               "type": "ASC"
+               }\`
+            </td>
+          </tr>
+        </tbody>
+      </Table>
+      "
+    `);
   });
 });