+++ processor/transform/compatability.ts
@@ -1,0 +1,58 @@
+import { Emphasis, Image, Strong, Node, Parent } from 'mdast';
+import { Transform } from 'mdast-util-from-markdown';
+import { phrasing } from 'mdast-util-phrasing';
+import { visit } from 'unist-util-visit';
+
+const strongTest = (node: Node): node is Strong | Emphasis => ['emphasis', 'strong'].includes(node.type);
+
+const addSpaceBefore = (index: number, parent: Parent) => {
+  if (!(index > 0 && parent.children[index - 1])) return;
+
+  const prev = parent.children[index - 1];
+  if (!('value' in prev) || prev.value.endsWith(' ') || prev.type === 'escape') return;
+
+  parent.children.splice(index, 0, { type: 'text', value: ' ' });
+};
+
+const addSpaceAfter = (index: number, parent: Parent) => {
+  if (!(index < parent.children.length - 1 && parent.children[index + 1])) return;
+
+  const nextChild = parent.children[index + 1];
+  if (!('value' in nextChild) || nextChild.value.startsWith(' ')) return;
+
+  parent.children.splice(index + 1, 0, { type: 'text', value: ' ' });
+};
+
+const trimEmphasis = (node: Emphasis | Strong, index: number, parent: Parent) => {
+  let trimmed = false;
+
+  visit(node, 'text', child => {
+    const newValue = child.value.trim();
+
+    if (newValue !== child.value) {
+      trimmed = true;
+      child.value = newValue;
+    }
+  });
+
+  if (trimmed) {
+    addSpaceBefore(index, parent);
+    addSpaceAfter(index, parent);
+  }
+
+  return node;
+};
+
+const compatibilityTransfomer = (): Transform => tree => {
+  visit(tree, strongTest, trimEmphasis);
+
+  visit(tree, 'image', (node: Image, index: number, parent: Parent) => {
+    if (phrasing(parent) || !parent.children.every(child => child.type === 'image' || !phrasing(child))) return;
+
+    parent.children.splice(index, 1, { type: 'paragraph', children: [node] });
+  });
+
+  return tree;
+};
+
+export default compatibilityTransfomer;